Unify Gerber and IPC geometry IR
