+++ crates/gerberx2/src/geometry/mod.rs
@@ -1,9 +1,3 @@
-pub mod compare;
 pub mod extract;
-pub mod ir;
-pub mod process;
-pub mod raster;
-pub mod svg;
-pub mod terminal;
 
 pub use extract::extract_document;
