+++ crates/gerberx2/src/parse.rs
@@ -1,5 +1,6 @@
 use crate::types::*;
 use crate::{GerberError, GerberX2, Interner, Result, Symbol};
+use pcb_ir::dialects::gerber::{Mirroring, Polarity};
 use std::collections::HashMap;
 
 pub struct Parser<'a> {
