+++ crates/pcb-ipc2581-tools/Cargo.toml
@@ -8,24 +8,20 @@ authors.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-base64 = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
 colored = { workspace = true }
 comfy-table = { workspace = true }
-i_overlay = { workspace = true }
 ipc2581 = { workspace = true }
 gerberx2 = { workspace = true }
 jiff = { workspace = true }
-kurbo = { workspace = true }
 log = { workspace = true }
 minijinja = { workspace = true }
+pcb-ir = { workspace = true }
 pcb-sch = { workspace = true, features = ["table"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
 quick-xml = { workspace = true }
-resvg = { workspace = true }
-terminal_size = { workspace = true }
 zstd = { workspace = true }
 pcb-diode-api = { workspace = true, optional = true }
 pcb-ui = { workspace = true, optional = true }
