--- crates/pcb-ipc2581-tools/src/geometry/svg.rs
@@ -1,743 +1,0 @@
-use std::fmt::Write;
-
-use super::ir::*;
-use ipc2581::types::LayerFunction;
-
-const VIEWBOX_PADDING_MM: f64 = 1.0;
-
-pub fn render_layer_bbox(doc: &GeometryDocument, layer_index: usize) -> BBox {
-    let layer = &doc.layers[layer_index];
-    let bbox = doc
-        .board_outlines
-        .iter()
-        .fold(layer.bbox, |bbox, outline| bbox.union(outline.bbox));
-
-    if bbox.is_empty() {
-        BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(100.0, 100.0),
-        }
-    } else if doc.board_outlines.is_empty() {
-        bbox
-    } else {
-        bbox.expand(VIEWBOX_PADDING_MM)
-    }
-}
-
-pub fn render_layer_svg(doc: &GeometryDocument, layer_index: usize) -> String {
-    render_layer_svg_with_size(doc, layer_index, None)
-}
-
-pub fn render_layer_svg_sized(
-    doc: &GeometryDocument,
-    layer_index: usize,
-    width_px: u32,
-    height_px: u32,
-) -> String {
-    render_layer_svg_with_size(doc, layer_index, Some((width_px, height_px)))
-}
-
-fn render_layer_svg_with_size(
-    doc: &GeometryDocument,
-    layer_index: usize,
-    pixel_size: Option<(u32, u32)>,
-) -> String {
-    let layer = &doc.layers[layer_index];
-    let geometry_bbox = render_layer_bbox(doc, layer_index);
-    let viewbox_y = -geometry_bbox.max.y;
-
-    let mut svg = String::new();
-    if let Some((width_px, height_px)) = pixel_size {
-        writeln!(
-            svg,
-            "<svg xmlns='http://www.w3.org/2000/svg' width='{width_px}' height='{height_px}' viewBox='{} {} {} {}'>",
-            fmt_num(geometry_bbox.min.x),
-            fmt_num(viewbox_y),
-            fmt_num(geometry_bbox.width()),
-            fmt_num(geometry_bbox.height())
-        )
-        .unwrap();
-    } else {
-        writeln!(
-            svg,
-            "<svg xmlns='http://www.w3.org/2000/svg' viewBox='{} {} {} {}'>",
-            fmt_num(geometry_bbox.min.x),
-            fmt_num(viewbox_y),
-            fmt_num(geometry_bbox.width()),
-            fmt_num(geometry_bbox.height())
-        )
-        .unwrap();
-    }
-    writeln!(svg, "  <title>{}</title>", escape_xml(&doc.board_name)).unwrap();
-    writeln!(
-        svg,
-        "  <g data-layer='{}' transform='scale(1 -1)'>",
-        escape_xml(&layer.name)
-    )
-    .unwrap();
-
-    write_feature_paths_by_bucket(&mut svg, doc, layer, &[FeatureBucket::Fill]);
-    write_feature_paths_by_bucket(&mut svg, doc, layer, &[FeatureBucket::Trace]);
-    write_feature_paths_by_bucket(&mut svg, doc, layer, &[FeatureBucket::Smd]);
-    write_feature_paths_by_bucket(&mut svg, doc, layer, &[FeatureBucket::Thermal]);
-    write_feature_paths_by_bucket(
-        &mut svg,
-        doc,
-        layer,
-        &[
-            FeatureBucket::Pth,
-            FeatureBucket::Via,
-            FeatureBucket::Cutout,
-            FeatureBucket::Antipad,
-        ],
-    );
-
-    write_board_outlines(&mut svg, doc);
-
-    writeln!(svg, "  </g>").unwrap();
-    writeln!(svg, "</svg>").unwrap();
-    svg
-}
-
-fn write_board_outlines(svg: &mut String, doc: &GeometryDocument) {
-    for outline in &doc.board_outlines {
-        for path in &doc.paths
-            [outline.path_start as usize..(outline.path_start + outline.path_count) as usize]
-        {
-            writeln!(
-                svg,
-                "    <path d='{}' fill='none' stroke='#000000' stroke-width='{}' stroke-linecap='{}' stroke-linejoin='round' data-board-outline='true'/>",
-                path_data(doc, path),
-                fmt_num(path.stroke_width),
-                line_cap(path.line_cap)
-            )
-            .unwrap();
-        }
-    }
-}
-
-fn write_feature_paths_by_bucket(
-    svg: &mut String,
-    doc: &GeometryDocument,
-    layer: &GeometryLayer,
-    buckets: &[FeatureBucket],
-) {
-    for feature in &doc.features
-        [layer.feature_start as usize..(layer.feature_start + layer.feature_count) as usize]
-    {
-        if !buckets.contains(&feature.bucket) {
-            continue;
-        }
-
-        for path in &doc.paths
-            [feature.path_start as usize..(feature.path_start + feature.path_count) as usize]
-        {
-            write_path(svg, doc, layer, feature, path);
-        }
-    }
-}
-
-fn write_path(
-    svg: &mut String,
-    doc: &GeometryDocument,
-    layer: &GeometryLayer,
-    feature: &GeometryFeature,
-    path: &GeometryPath,
-) {
-    let d = path_data(doc, path);
-    let style = style_for_feature(layer, feature);
-    let fill_rule = match path.fill_rule {
-        FillRule::NonZero => "nonzero",
-        FillRule::EvenOdd => "evenodd",
-    };
-
-    if path.flags.stroked {
-        writeln!(
-            svg,
-            "    <path d='{}' fill='none' stroke='{}' stroke-opacity='{}' stroke-width='{}' stroke-linecap='{}' stroke-linejoin='round'/>",
-            d,
-            style.color,
-            fmt_num(style.opacity),
-            fmt_num(path.stroke_width),
-            line_cap(path.line_cap)
-        )
-        .unwrap();
-    } else if path.flags.filled {
-        writeln!(
-            svg,
-            "    <path d='{}' fill='{}' fill-opacity='{}' fill-rule='{}'/>",
-            d,
-            style.color,
-            fmt_num(style.opacity),
-            fill_rule
-        )
-        .unwrap();
-    }
-}
-
-fn path_data(doc: &GeometryDocument, path: &GeometryPath) -> String {
-    let mut data = String::new();
-    for contour in &doc.contours
-        [path.contour_start as usize..(path.contour_start + path.contour_count) as usize]
-    {
-        let mut current = Point::default();
-        for cmd in &doc.path_cmds
-            [contour.cmd_start as usize..(contour.cmd_start + contour.cmd_count) as usize]
-        {
-            match cmd.op {
-                PathOp::MoveTo => {
-                    current = cmd.p0;
-                    if !data.is_empty() {
-                        data.push(' ');
-                    }
-                    write!(data, "M{} {}", fmt_num(cmd.p0.x), fmt_num(cmd.p0.y)).unwrap()
-                }
-                PathOp::LineTo => {
-                    current = cmd.p0;
-                    write!(data, " L{} {}", fmt_num(cmd.p0.x), fmt_num(cmd.p0.y)).unwrap()
-                }
-                PathOp::ArcTo => {
-                    let radius = current.distance_to(cmd.p1);
-                    let sweep = if cmd.clockwise { 0 } else { 1 };
-                    if current.distance_to(cmd.p0) <= 1e-9 && radius > 0.0 {
-                        let mid =
-                            Point::new(2.0 * cmd.p1.x - current.x, 2.0 * cmd.p1.y - current.y);
-                        write!(
-                            data,
-                            " A{} {} 0 0 {} {} {} A{} {} 0 0 {} {} {}",
-                            fmt_num(radius),
-                            fmt_num(radius),
-                            sweep,
-                            fmt_num(mid.x),
-                            fmt_num(mid.y),
-                            fmt_num(radius),
-                            fmt_num(radius),
-                            sweep,
-                            fmt_num(cmd.p0.x),
-                            fmt_num(cmd.p0.y)
-                        )
-                        .unwrap();
-                        current = cmd.p0;
-                        continue;
-                    }
-
-                    let large_arc = u8::from(
-                        arc_sweep_radians(current, cmd.p0, cmd.p1, cmd.clockwise)
-                            > std::f64::consts::PI,
-                    );
-                    current = cmd.p0;
-                    write!(
-                        data,
-                        " A{} {} 0 {} {} {} {}",
-                        fmt_num(radius),
-                        fmt_num(radius),
-                        large_arc,
-                        sweep,
-                        fmt_num(cmd.p0.x),
-                        fmt_num(cmd.p0.y)
-                    )
-                    .unwrap()
-                }
-                PathOp::CubicTo => write!(
-                    data,
-                    " C{} {},{} {},{} {}",
-                    fmt_num(cmd.p0.x),
-                    fmt_num(cmd.p0.y),
-                    fmt_num(cmd.p1.x),
-                    fmt_num(cmd.p1.y),
-                    fmt_num(cmd.p2.x),
-                    fmt_num(cmd.p2.y)
-                )
-                .map(|_| current = cmd.p2)
-                .unwrap(),
-                PathOp::Close => data.push_str(" Z"),
-            }
-        }
-    }
-    data
-}
-
-#[derive(Debug, Clone, Copy)]
-struct Style {
-    color: &'static str,
-    opacity: f64,
-}
-
-fn style_for_feature(layer: &GeometryLayer, feature: &GeometryFeature) -> Style {
-    if feature.kind == FeatureKind::FlattenedBucket
-        && feature.polarity == GeometryPolarity::Positive
-    {
-        return flat_layer_style(layer);
-    }
-    Style {
-        color: color_for_bucket(feature.bucket, feature.polarity),
-        opacity: opacity_for_bucket(feature.bucket, feature.polarity),
-    }
-}
-
-fn flat_layer_style(layer: &GeometryLayer) -> Style {
-    match layer.layer_function {
-        LayerFunction::Conductor
-        | LayerFunction::CondFilm
-        | LayerFunction::CondFoil
-        | LayerFunction::Plane
-        | LayerFunction::Signal
-        | LayerFunction::Mixed => Style {
-            color: "#d87822",
-            opacity: 0.9,
-        },
-        LayerFunction::Solderpaste | LayerFunction::Pastemask => Style {
-            color: "#aeb4bb",
-            opacity: 0.9,
-        },
-        LayerFunction::Soldermask => Style {
-            color: "#159447",
-            opacity: 0.55,
-        },
-        LayerFunction::Silkscreen | LayerFunction::Legend => Style {
-            color: "#f8f9fa",
-            opacity: 0.95,
-        },
-        LayerFunction::Rout | LayerFunction::BoardOutline => Style {
-            color: "#606060",
-            opacity: 1.0,
-        },
-        _ => Style {
-            color: "#5c7cfa",
-            opacity: 0.85,
-        },
-    }
-}
-
-fn color_for_bucket(bucket: FeatureBucket, polarity: GeometryPolarity) -> &'static str {
-    if polarity == GeometryPolarity::Negative {
-        return "#111827";
-    }
-
-    match bucket {
-        FeatureBucket::Smd => "#d97706",
-        FeatureBucket::Pth => "#7c3aed",
-        FeatureBucket::Via => "#2563eb",
-        FeatureBucket::Trace => "#c2410c",
-        FeatureBucket::Fill => "#2f9e44",
-        FeatureBucket::Cutout => "#64748b",
-        FeatureBucket::Thermal => "#ca8a04",
-        FeatureBucket::Antipad => "#64748b",
-    }
-}
-
-fn opacity_for_bucket(bucket: FeatureBucket, polarity: GeometryPolarity) -> f64 {
-    if polarity == GeometryPolarity::Negative {
-        return 0.8;
-    }
-
-    match bucket {
-        FeatureBucket::Fill => 0.78,
-        FeatureBucket::Trace => 0.95,
-        FeatureBucket::Smd | FeatureBucket::Pth => 0.95,
-        FeatureBucket::Via => 0.9,
-        FeatureBucket::Cutout | FeatureBucket::Antipad => 1.0,
-        FeatureBucket::Thermal => 0.92,
-    }
-}
-
-fn line_cap(line_cap: LineCap) -> &'static str {
-    match line_cap {
-        LineCap::Round => "round",
-        LineCap::Square => "square",
-        LineCap::Butt => "butt",
-    }
-}
-
-fn fmt_num(value: f64) -> String {
-    let value = if value.abs() < 0.000_000_5 {
-        0.0
-    } else {
-        value
-    };
-    format!("{value:.6}")
-        .trim_end_matches('0')
-        .trim_end_matches('.')
-        .to_string()
-}
-
-fn escape_xml(value: &str) -> String {
-    value
-        .replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&apos;")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn renders_arc_path_command_without_flattening() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let mut bbox = BBox::empty();
-        bbox.include_point(Point::new(-1.0, 0.0));
-        bbox.include_point(Point::new(1.0, 0.0));
-        bbox.include_point(Point::new(0.0, -1.0));
-        doc.push_path(
-            GeometryPath::filled(FillRule::NonZero, bbox),
-            [
-                PathCmd::move_to(Point::new(1.0, 0.0)),
-                PathCmd::arc_to(Point::new(-1.0, 0.0), Point::new(0.0, 0.0), true),
-                PathCmd::close(),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_count: 1,
-            bbox,
-            ..GeometryFeature::new(
-                FeatureKind::Polygon,
-                FeatureBucket::Fill,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 1,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert!(svg.contains(" A1 1 0 0 0 -1 0"));
-    }
-
-    #[test]
-    fn renders_full_circle_arc_as_two_svg_arcs() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(-1.0, -1.0),
-            max: Point::new(1.0, 1.0),
-        };
-        doc.push_path(
-            GeometryPath::filled(FillRule::NonZero, bbox),
-            [
-                PathCmd::move_to(Point::new(1.0, 0.0)),
-                PathCmd::arc_to(Point::new(1.0, 0.0), Point::new(0.0, 0.0), false),
-                PathCmd::close(),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_count: 1,
-            bbox,
-            ..GeometryFeature::new(
-                FeatureKind::Polygon,
-                FeatureBucket::Fill,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 1,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert!(svg.contains(" A1 1 0 0 1 -1 0 A1 1 0 0 1 1 0"));
-    }
-
-    #[test]
-    fn renders_compound_path_contours_as_one_svg_path() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let outer = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(10.0, 10.0),
-        };
-        let inner = BBox {
-            min: Point::new(2.0, 2.0),
-            max: Point::new(4.0, 4.0),
-        };
-        doc.push_compound_path(
-            GeometryPath::filled(FillRule::EvenOdd, BBox::empty()),
-            [
-                (
-                    outer,
-                    vec![
-                        PathCmd::move_to(Point::new(0.0, 0.0)),
-                        PathCmd::line_to(Point::new(10.0, 0.0)),
-                        PathCmd::line_to(Point::new(10.0, 10.0)),
-                        PathCmd::line_to(Point::new(0.0, 10.0)),
-                        PathCmd::close(),
-                    ],
-                ),
-                (
-                    inner,
-                    vec![
-                        PathCmd::move_to(Point::new(2.0, 2.0)),
-                        PathCmd::line_to(Point::new(4.0, 2.0)),
-                        PathCmd::line_to(Point::new(4.0, 4.0)),
-                        PathCmd::line_to(Point::new(2.0, 4.0)),
-                        PathCmd::close(),
-                    ],
-                ),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_count: 1,
-            bbox: outer,
-            ..GeometryFeature::new(
-                FeatureKind::Polygon,
-                FeatureBucket::Fill,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 1,
-            bbox: outer,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert_eq!(svg.matches("<path d='").count(), 1);
-        assert!(svg.contains(" Z M2 2"));
-        assert!(svg.contains("fill-rule='evenodd'"));
-    }
-
-    #[test]
-    fn renders_cutout_features_after_positive_geometry() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(1.0, 1.0),
-        };
-        doc.push_path(
-            GeometryPath::filled(FillRule::NonZero, bbox),
-            [
-                PathCmd::move_to(Point::new(0.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 1.0)),
-                PathCmd::close(),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_count: 1,
-            bbox,
-            ..GeometryFeature::new(
-                FeatureKind::Polygon,
-                FeatureBucket::Fill,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.push_path(
-            GeometryPath::filled(FillRule::NonZero, bbox),
-            [
-                PathCmd::move_to(Point::new(0.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 1.0)),
-                PathCmd::close(),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_start: 1,
-            path_count: 1,
-            bbox,
-            ..GeometryFeature::new(
-                FeatureKind::Hole,
-                FeatureBucket::Cutout,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 2,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert_eq!(svg.matches("<path d='").count(), 2);
-        assert!(svg.find("fill='#2f9e44'").unwrap() < svg.find("fill='#64748b'").unwrap());
-    }
-
-    #[test]
-    fn renders_features_in_visual_stack_order() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(1.0, 1.0),
-        };
-
-        for (kind, bucket) in [
-            (FeatureKind::Hole, FeatureBucket::Cutout),
-            (FeatureKind::Padstack, FeatureBucket::Smd),
-            (FeatureKind::Trace, FeatureBucket::Trace),
-            (FeatureKind::Polygon, FeatureBucket::Fill),
-        ] {
-            let path_start = doc.paths.len() as u32;
-            doc.push_path(
-                GeometryPath::filled(FillRule::NonZero, bbox),
-                [
-                    PathCmd::move_to(Point::new(0.0, 0.0)),
-                    PathCmd::line_to(Point::new(1.0, 0.0)),
-                    PathCmd::line_to(Point::new(1.0, 1.0)),
-                    PathCmd::close(),
-                ],
-            );
-            doc.features.push(GeometryFeature {
-                path_start,
-                path_count: 1,
-                bbox,
-                ..GeometryFeature::new(kind, bucket, GeometryPolarity::Positive)
-            });
-        }
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 4,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        let fill = svg.find("fill='#2f9e44'").unwrap();
-        let trace = svg.find("fill='#c2410c'").unwrap();
-        let smd = svg.find("fill='#d97706'").unwrap();
-        let cutout = svg.find("fill='#64748b'").unwrap();
-        assert!(fill < trace);
-        assert!(trace < smd);
-        assert!(smd < cutout);
-        assert!(svg.contains("fill='#2f9e44' fill-opacity='0.78'"));
-        assert!(svg.contains("fill='#64748b' fill-opacity='1'"));
-    }
-
-    #[test]
-    fn renders_flattened_masks_with_gerber_layer_colors() {
-        assert_flat_color("F.Cu", LayerFunction::Signal, "#d87822", "0.9");
-        assert_flat_color("F.Paste", LayerFunction::Solderpaste, "#aeb4bb", "0.9");
-        assert_flat_color("F.Mask", LayerFunction::Soldermask, "#159447", "0.55");
-        assert_flat_color("F.SilkS", LayerFunction::Silkscreen, "#f8f9fa", "0.95");
-        assert_flat_color("LAMINATE", LayerFunction::DielCore, "#5c7cfa", "0.85");
-    }
-
-    fn assert_flat_color(
-        layer_name: &str,
-        layer_function: LayerFunction,
-        color: &str,
-        opacity: &str,
-    ) {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(1.0, 1.0),
-        };
-        doc.push_path(
-            GeometryPath::filled(FillRule::NonZero, bbox),
-            [
-                PathCmd::move_to(Point::new(0.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 0.0)),
-                PathCmd::line_to(Point::new(1.0, 1.0)),
-                PathCmd::close(),
-            ],
-        );
-        doc.features.push(GeometryFeature {
-            path_count: 1,
-            bbox,
-            ..GeometryFeature::new(
-                FeatureKind::FlattenedBucket,
-                FeatureBucket::Fill,
-                GeometryPolarity::Positive,
-            )
-        });
-        doc.layers.push(GeometryLayer {
-            name: layer_name.to_string(),
-            source_layer_ref: interner.intern(layer_name),
-            layer_function,
-            feature_start: 0,
-            feature_count: 1,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert!(svg.contains(&format!("fill='{color}' fill-opacity='{opacity}'")));
-    }
-
-    #[test]
-    fn renders_layer_without_outline_without_viewbox_padding() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(10.0, 5.0),
-        };
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 0,
-            bbox,
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert!(svg.contains("viewBox='0 -5 10 5'"));
-    }
-
-    #[test]
-    fn renders_board_outline_as_black_overlay_and_includes_it_in_viewbox() {
-        let mut interner = ipc2581::Interner::new();
-        let mut doc = GeometryDocument::new("test".to_string());
-        let bbox = BBox {
-            min: Point::new(0.0, 0.0),
-            max: Point::new(10.0, 5.0),
-        };
-        doc.push_path(
-            GeometryPath::stroked(0.1, LineCap::Round, bbox),
-            [
-                PathCmd::move_to(Point::new(0.0, 0.0)),
-                PathCmd::line_to(Point::new(10.0, 0.0)),
-                PathCmd::line_to(Point::new(10.0, 5.0)),
-                PathCmd::close(),
-            ],
-        );
-        doc.board_outlines.push(BoardOutline {
-            path_start: 0,
-            path_count: 1,
-            bbox,
-        });
-        doc.layers.push(GeometryLayer {
-            name: "F.Cu".to_string(),
-            source_layer_ref: interner.intern("F.Cu"),
-            layer_function: ipc2581::types::LayerFunction::Signal,
-            feature_start: 0,
-            feature_count: 0,
-            bbox: BBox::empty(),
-        });
-
-        let svg = render_layer_svg(&doc, 0);
-
-        assert!(svg.contains("viewBox='-1 -6 12 7'"));
-        assert!(svg.contains("stroke='#000000'"));
-        assert!(svg.contains("data-board-outline='true'"));
-    }
-}
