+++ crates/pcb-ir/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pcb-ir"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+
+[dependencies]
+base64 = { workspace = true }
+i_overlay = { workspace = true }
+kurbo = { workspace = true }
+resvg = { workspace = true }
+terminal_size = { workspace = true }
