+++ crates/pcb-ir/src/dialects/gerber/terminal.rs
@@ -1,11 +1,10 @@
 use std::io::{self, IsTerminal, Write};
 
-use anyhow::{Result, bail};
 use base64::Engine;
 use base64::engine::general_purpose::STANDARD;
 use terminal_size::{Width, terminal_size};
 
-use super::ir::GeometryDocument;
+use crate::dialects::gerber::GeometryDocument;
 
 const KITTY_CHUNK_SIZE: usize = 4096;
 const MAX_TERMINAL_DIMENSION_PX: u32 = 1200;
@@ -14,19 +13,16 @@ pub fn can_render_to_terminal() -> bool {
     io::stdout().is_terminal()
 }
 
-pub fn render_layer_to_terminal(doc: &GeometryDocument, layer_index: usize) -> Result<()> {
+pub fn render_to_terminal<A>(doc: &GeometryDocument<A>) -> Result<(), String> {
     if !io::stdout().is_terminal() {
-        bail!("stdout is not an interactive terminal; pass --output <path>.svg or <path>.png");
+        return Err(
+            "stdout is not an interactive terminal; pass an SVG or PNG output path".to_string(),
+        );
     }
-
-    let png = super::raster::render_layer_png_with_max_dimension(
-        doc,
-        layer_index,
-        terminal_max_dimension_px(),
-    )?;
+    let png = super::raster::render_png_with_max_dimension(doc, terminal_max_dimension_px())?;
     let mut stdout = io::stdout().lock();
-    write_kitty_png(&mut stdout, &png)?;
-    stdout.write_all(b"\n")?;
+    write_kitty_png(&mut stdout, &png).map_err(|err| err.to_string())?;
+    stdout.write_all(b"\n").map_err(|err| err.to_string())?;
     Ok(())
 }
 
@@ -34,17 +30,15 @@ fn terminal_max_dimension_px() -> u32 {
     let Some((Width(columns), _)) = terminal_size() else {
         return MAX_TERMINAL_DIMENSION_PX;
     };
-
     u32::from(columns)
         .saturating_mul(12)
         .clamp(1, MAX_TERMINAL_DIMENSION_PX)
 }
 
-fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
+pub fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
     let encoded = STANDARD.encode(png);
     let mut chunks = encoded.as_bytes().chunks(KITTY_CHUNK_SIZE).peekable();
     let mut first = true;
-
     while let Some(chunk) = chunks.next() {
         let more = u8::from(chunks.peek().is_some());
         if first {
@@ -56,7 +50,6 @@ fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
         writer.write_all(chunk)?;
         writer.write_all(b"\x1b\\")?;
     }
-
     Ok(())
 }
