+++ crates/pcb-ir/src/dialects/ipc/terminal.rs
@@ -4,7 +4,7 @@ use base64::Engine;
 use base64::engine::general_purpose::STANDARD;
 use terminal_size::{Width, terminal_size};
 
-use super::ir::GeometryDocument;
+use crate::dialects::ipc::GeometryDocument;
 
 const KITTY_CHUNK_SIZE: usize = 4096;
 const MAX_TERMINAL_DIMENSION_PX: u32 = 1200;
@@ -13,16 +13,30 @@ pub fn can_render_to_terminal() -> bool {
     io::stdout().is_terminal()
 }
 
-pub fn render_to_terminal(doc: &GeometryDocument) -> crate::Result<()> {
+pub fn render_layer_to_terminal<S, L, F>(
+    doc: &GeometryDocument<S, L>,
+    layer_index: usize,
+    flat_style: F,
+) -> Result<(), String>
+where
+    F: Copy + Fn(&L) -> (&'static str, f64),
+{
     if !io::stdout().is_terminal() {
-        return Err(crate::GerberError::Render(
-            "stdout is not an interactive terminal; pass an SVG or PNG output path".to_string(),
-        ));
+        return Err(
+            "stdout is not an interactive terminal; pass --output <path>.svg or <path>.png"
+                .to_string(),
+        );
     }
-    let png = super::raster::render_png_with_max_dimension(doc, terminal_max_dimension_px())?;
+
+    let png = super::raster::render_layer_png_with_max_dimension(
+        doc,
+        layer_index,
+        terminal_max_dimension_px(),
+        flat_style,
+    )?;
     let mut stdout = io::stdout().lock();
-    write_kitty_png(&mut stdout, &png)?;
-    stdout.write_all(b"\n")?;
+    write_kitty_png(&mut stdout, &png).map_err(|err| err.to_string())?;
+    stdout.write_all(b"\n").map_err(|err| err.to_string())?;
     Ok(())
 }
 
@@ -30,15 +44,17 @@ fn terminal_max_dimension_px() -> u32 {
     let Some((Width(columns), _)) = terminal_size() else {
         return MAX_TERMINAL_DIMENSION_PX;
     };
+
     u32::from(columns)
         .saturating_mul(12)
         .clamp(1, MAX_TERMINAL_DIMENSION_PX)
 }
 
-fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
+pub fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
     let encoded = STANDARD.encode(png);
     let mut chunks = encoded.as_bytes().chunks(KITTY_CHUNK_SIZE).peekable();
     let mut first = true;
+
     while let Some(chunk) = chunks.next() {
         let more = u8::from(chunks.peek().is_some());
         if first {
@@ -50,5 +66,6 @@ fn write_kitty_png<W: Write>(writer: &mut W, png: &[u8]) -> io::Result<()> {
         writer.write_all(chunk)?;
         writer.write_all(b"\x1b\\")?;
     }
+
     Ok(())
 }
