+++ crates/pcb-ir/src/dialects/mod.rs
@@ -1,0 +1,2 @@
+pub mod gerber;
+pub mod ipc;
