+++ crates/pcb-ir/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod common;
+pub mod dialects;
