+++ crates/pcb/Cargo.toml
@@ -50,6 +50,7 @@ inquire = { workspace = true }
 pcb-ui = { workspace = true }
 pcb-fmt = { workspace = true }
 gerberx2 = { workspace = true }
+pcb-ir = { workspace = true }
 similar = { workspace = true }
 pathdiff = { workspace = true }
 petgraph = { workspace = true }
